edited Internationalization and double to BigDecimal

+++ service/src/main/java/by/work/service/OrderService.java
@@ -11,5 +11,5 @@ public interface OrderService {
 
     Order save(ProductDTO products);
     BigDecimal getTotalAmount(Set<Product> products);
-    void update(Order order,Double totalAmount);
+    void update(Order order, BigDecimal totalAmount);
 }
