+++ service/src/main/java/by/work/service/dto/OrderDTO.java
@@ -2,11 +2,12 @@ package by.work.service.dto;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.List;
 
 @Data
 public class OrderDTO {
-    List<Double> price = new ArrayList<>();
+    List<BigDecimal> price = new ArrayList<>();
     List<Integer> count = new ArrayList<>();
 }
